Implement a simple output plugin for NATS servers.
Basically this is a thin wrapper over the ruby client for NATS.
Here is how it work:

  1. Create a separate thread for running NATS client
  2. Events are passed to the NATS client by calling `NATS.publish`.
     (This is done inside `EM.next_tick`)
  3. Call NATS.stop on shutdown.

The beauty of this approach is that most of the tedious work (like
connection management and error handling) is managed by the NATS
client transparently. So we do not need to reinvent the wheel for
ourselves.